task 0 solvedwq

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -7,7 +7,7 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	int umubare = 0;
+	size_t umubare = 0;
 
 	while (h != NULL)
 	{
